@@ -2,16 +2,13 @@ package kr.cds.jisulife;
 
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-
 import kr.go.seoul.airquality.AirQualityButtonTypeA;
-import kr.go.seoul.airquality.AirQualityButtonTypeB;
-import kr.go.seoul.airquality.AirQualityMapTypeA;
 import kr.go.seoul.airquality.AirQualityTypeMini;
 
 
