@@ -4,35 +4,31 @@ import android.content.Context;
 import android.graphics.Color;
 import android.graphics.PorterDuff;
 import android.os.Bundle;
-import android.support.design.widget.NavigationView;
-import android.support.v4.app.FragmentTransaction;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import com.google.android.material.navigation.NavigationView;
+import androidx.fragment.app.FragmentTransaction;
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.inputmethod.InputMethodManager;
 
-import com.google.android.gms.ads.AdRequest;
-import com.google.android.gms.ads.AdView;
-import com.google.android.gms.ads.MobileAds;
-import com.tsengvn.typekit.TypekitContextWrapper;
-
 public class MainActivity extends AppCompatActivity
-        implements NavigationView.OnNavigationItemSelectedListener, CommunicationListener{
+        implements NavigationView.OnNavigationItemSelectedListener, CommunicationListener {
 
     public static DbHelper dbHelper;
     public String location = "1100000000";
     WeatherFragment weatherFragment;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
         dbHelper = DbHelper.getsInstance(this);
-        if(savedInstanceState == null){
+        if (savedInstanceState == null) {
             dbHelper = DbHelper.getsInstance(this);
         }
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
@@ -49,7 +45,7 @@ public class MainActivity extends AppCompatActivity
             public void onDrawerOpened(View drawerView) {
                 InputMethodManager inputMethodManager = (InputMethodManager)
                         getSystemService(Context.INPUT_METHOD_SERVICE);
-                inputMethodManager.hideSoftInputFromWindow(getCurrentFocus().getWindowToken(), 0);
+                inputMethodManager.hideSoftInputFromWindow(drawerView.getWindowToken(), 0);
 
             }
 
@@ -71,14 +67,10 @@ public class MainActivity extends AppCompatActivity
 
         NavigationView navigationView = (NavigationView) findViewById(R.id.nav_view);
         navigationView.setNavigationItemSelectedListener(this);
-              getSupportActionBar().setTitle("");
+        getSupportActionBar().setTitle("");
 
-        MobileAds.initialize(this, "ca-app-pub-4712412246131543~2639273924");
-        AdView mAdView = (AdView)this.findViewById(R.id.adView);
+        //MobileAds.initialize(this, "ca-app-pub-4712412246131543~2639273924");
 
-         AdRequest adRequest = new AdRequest.Builder().build();
-
-         mAdView.loadAd(adRequest);
         FragmentTransaction transaction = getSupportFragmentManager().beginTransaction();   //최초화면
 
         weatherFragment = new WeatherFragment();
@@ -113,14 +105,11 @@ public class MainActivity extends AppCompatActivity
         int id = item.getItemId();
         if (id == R.id.nav_weather) {
             transaction.replace(R.id.content_frame, weatherFragment);
-        }
-        else if (id == R.id.nav_mosquito) {
+        } else if (id == R.id.nav_mosquito) {
             transaction.replace(R.id.content_frame, mosquitoFragment);
-        }
-        else if (id == R.id.nav_expression) {
+        } else if (id == R.id.nav_expression) {
             transaction.replace(R.id.content_frame, expressionFragment);
-        }
-        else if (id == R.id.nav_air) {
+        } else if (id == R.id.nav_air) {
             transaction.replace(R.id.content_frame, airFragment);
         }
         else if (id == R.id.nav_location) {
@@ -131,11 +120,6 @@ public class MainActivity extends AppCompatActivity
         DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
         drawer.closeDrawer(GravityCompat.START);
         return true;
-    }
-
-    @Override
-    protected void attachBaseContext(Context newBase) {
-        super.attachBaseContext(TypekitContextWrapper.wrap(newBase));
     }
 
     @Override
