@@ -8,9 +8,9 @@ import android.graphics.drawable.GradientDrawable;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuInflater;
@@ -19,6 +19,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.view.animation.Animation;
 import android.view.animation.TranslateAnimation;
+import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.ProgressBar;
 import android.widget.ScrollView;
@@ -27,8 +28,10 @@ import android.widget.Toast;
 
 import com.kakao.kakaolink.KakaoLink;
 import com.kakao.kakaolink.KakaoTalkLinkMessageBuilder;
+
 import org.xmlpull.v1.XmlPullParser;
 import org.xmlpull.v1.XmlPullParserFactory;
+
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.net.URL;
@@ -37,13 +40,16 @@ import java.util.Calendar;
 import java.util.Date;
 import java.util.Locale;
 
-public class MosquitoFragment extends Fragment{
+public class MosquitoFragment extends Fragment {
     private boolean INTERNET_STATE = true;
     TextView textView2;
     TextView textView3;
     TextView textView4;
     TextView textView12;
     TextView textView13;
+    Button btnHouse;
+    Button btnWater;
+    Button btnPark;
     private String value;   //모기지수
     ImageView imageView;
     ScrollView scrollView;
@@ -51,19 +57,17 @@ public class MosquitoFragment extends Fragment{
     int textColor = 0;
     String date;
     String grade;
-    public static mosquitoEntry mosquitoEntry;
+    public static MosquitoEntry mosquitoEntry;
     private ProgressBar progressBar;
 
-    public void shareKakao()
-    {
-        try{
+    public void shareKakao() {
+        try {
             final KakaoLink kakaoLink = KakaoLink.getKakaoLink(getActivity());
             final KakaoTalkLinkMessageBuilder kakaoBuilder = kakaoLink.createKakaoTalkLinkMessageBuilder();
-            kakaoBuilder.addText("오늘의 모기지수는 "+grade+" "+value+"입니다.");
+            kakaoBuilder.addText("오늘의 모기지수는 " + grade + " " + value + "입니다.");
             kakaoBuilder.addAppButton("앱 실행 혹은 다운로드");
             kakaoLink.sendMessage(kakaoBuilder, getContext());
-        }catch (Exception e)
-        {
+        } catch (Exception e) {
             e.printStackTrace();
         }
     }
@@ -73,10 +77,11 @@ public class MosquitoFragment extends Fragment{
         inflater.inflate(R.menu.mosquito_actions, menu);
         super.onCreateOptionsMenu(menu, inflater);
     }
+
     public boolean onOptionsItemSelected(MenuItem item) {
         if (item.getItemId() == R.id.refresh) {
-            INTERNET_STATE=checkInternet();
-            thread();
+            INTERNET_STATE = checkInternet();
+            thread("MOSQUITO_VALUE_HOUSE");
             return (true);
         }
         if (item.getItemId() == R.id.share) {
@@ -90,71 +95,96 @@ public class MosquitoFragment extends Fragment{
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        ((AppCompatActivity)getActivity()).getSupportActionBar().setTitle("모기예보");
-        ((AppCompatActivity)getActivity()).getSupportActionBar().setElevation(0);
+        ((AppCompatActivity) getActivity()).getSupportActionBar().setTitle("모기예보");
+        ((AppCompatActivity) getActivity()).getSupportActionBar().setElevation(0);
         return inflater.inflate(R.layout.fragment_mosquito, container, false);
 
     }
 
 
-    public void onViewCreated(View view, @Nullable Bundle savedInstanceState)
-    {
+    public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
 
         setRetainInstance(true);
         setHasOptionsMenu(true);
-        textView2 = (TextView) getView().findViewById(R.id.textView2);
-        textView3 = (TextView) getView().findViewById(R.id.textView3);
-        textView4 = (TextView) getView().findViewById(R.id.textView4);
-        textView12 = (TextView) getView().findViewById(R.id.textView12);
-        textView13 = (TextView) getView().findViewById(R.id.textView13);
-        imageView = (ImageView) getView().findViewById(R.id.indicator);
-        scrollView = (ScrollView) getView().findViewById(R.id.backGround);
+        textView2 = getView().findViewById(R.id.textView2);
+        textView3 = getView().findViewById(R.id.textView3);
+        textView4 = getView().findViewById(R.id.textView4);
+        textView12 = getView().findViewById(R.id.textView12);
+        textView13 = getView().findViewById(R.id.textView13);
+        imageView = getView().findViewById(R.id.indicator);
+        scrollView = getView().findViewById(R.id.backGround);
+        btnHouse = getView().findViewById(R.id.btnHouse);
+        btnPark = getView().findViewById(R.id.btnPark);
+        btnWater = getView().findViewById(R.id.btnWater);
+        scrollView = getView().findViewById(R.id.backGround);
         key1 = getResources().getString(R.string.key1);
-        progressBar = (ProgressBar) getView().findViewById(R.id.progressBar2);
+        progressBar = getView().findViewById(R.id.progressBar2);
         progressBar.setVisibility(View.GONE);
 
-        if(checkInternet()==false){
-            INTERNET_STATE=false;
-            if(getPreferences()==false)
-                Toast.makeText(getActivity(),"네트워크가 연결되지 않았습니다.", Toast.LENGTH_SHORT).show();
+        btnHouse.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                thread("MOSQUITO_VALUE_HOUSE");
+            }
+        });
 
+        btnWater.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                thread("MOSQUITO_VALUE_WATER");
+            }
+        });
+
+        btnPark.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                thread("MOSQUITO_VALUE_PARK");
+            }
+        });
+
+
+        if (checkInternet() == false) {
+            INTERNET_STATE = false;
+
+            if (getPreferences() == false)
+                Toast.makeText(getActivity(), "네트워크가 연결되지 않았습니다.", Toast.LENGTH_SHORT).show();
         }
-        thread();
+
+        thread("MOSQUITO_VALUE_HOUSE");
     }
 
-    public void thread(){
+    public void thread(final String criteria) {
         Thread t = new Thread(new Runnable() {
             @Override
             public void run() {
                 // TODO Auto-generated method stub
-                if(INTERNET_STATE) {
-                    value = getXmlData(getDateString()); //모기지수
-                }
-                else
+                if (INTERNET_STATE) {
+                    value = getXmlData(getDateString(), criteria); //모기지수
+                } else
                     return;
                 getActivity().runOnUiThread(new Runnable() {
                     @Override
                     public void run() {
                         progressBar.setVisibility(View.VISIBLE);
                         // TODO Auto-generated method stub
-                        date=getDateStringToShow();
+                        date = getDateStringToShow();
                         textView2.setText(date);
-                        textView3.setText(value);  //TextView에 문자열  value 출력
+                        textView3.setText(value);  //TextView에 문자열  valueTextView 출력
                         textView4.setText(getGrade(value));
 
                         mosquitoEntry = MainActivity.dbHelper.fetchEntryByIndex(getGradeDetail(Double.valueOf(value)));
-                        grade= mosquitoEntry.getmGrade();
+                        grade = mosquitoEntry.getmGrade();
 
                         textView12.setText(mosquitoEntry.getDefence_activity());
                         textView13.setText(mosquitoEntry.getAggressive_activity());
 
                         setColor();
                         TranslateAnimation anim = new TranslateAnimation
-                                (Animation.RELATIVE_TO_SELF,0,   // fromXDelta
-                                        Animation.RELATIVE_TO_SELF, Float.valueOf(value)/(float)37.2,  // toXDelta
+                                (Animation.RELATIVE_TO_SELF, 0,           // fromXDelta
+                                        Animation.RELATIVE_TO_SELF, Float.valueOf(value) < 120 ? (int)(Float.valueOf(value) / 4.3) : 26,  // toXDelta
                                         Animation.RELATIVE_TO_SELF, 0,    // fromYDelta
-                                        Animation.RELATIVE_TO_SELF, 0);// toYDelta
+                                        Animation.RELATIVE_TO_SELF, 0);     // toYDelta
                         anim.setDuration(1000);
                         anim.setFillAfter(true);
                         imageView.startAnimation(anim);
@@ -165,46 +195,46 @@ public class MosquitoFragment extends Fragment{
         });
         t.start();
     }
+
     private long getGradeDetail(double value) {
         //1단계
-        if(0<=value&&value<=500) {
-            if (0 <= value && value <= 83.3) {
+        if (0 <= value && value <= 60) {
+            if (0 <= value && value <= 10) {
                 return 1;
-            } else if (83.4 <= value && value <= 166.6) {
+            } else if (10.1 <= value && value <= 20) {
                 return 2;
-            } else if (166.7 <= value && value <= 250) {
+            } else if (20.1 <= value && value <= 30) {
                 return 3;
             }
 
             //2단계
-            else if (250.1 <= value && value <= 333.3) {
+            else if (30.1 <= value && value <= 40) {
                 return 4;
-            } else if (333.4 <= value && value <= 416.6) {
+            } else if (40.1 <= value && value <= 50) {
                 return 5;
-            } else if (416.7 <= value && value <= 500) {
+            } else if (50.1 <= value && value <= 60) {
                 return 6;
             }
-        }
-        else if(500.1<=value&&value<=1000) {
+        } else if (60.1 <= value && value <= 120) {
             //3단계
-            if (500.1 <= value && value <= 583.3) {
+            if (60.1 <= value && value <= 70) {
                 return 7;
-            } else if (583.4 <= value && value <= 666.6) {
+            } else if (70.1 <= value && value <= 80) {
                 return 8;
-            } else if (666.7 <= value && value <= 750) {
+            } else if (80.1 <= value && value <= 90) {
                 return 9;
             }
 
             //4단계
-            else if (750.1 <= value && value <= 833.3) {
+            else if (90.1 <= value && value <= 100) {
                 return 10;
-            } else if (833.4 <= value && value <= 916.6) {
+            } else if (100.1 <= value && value <= 110) {
                 return 11;
-            } else if (916.7 <= value && value <= 1000) {
+            } else if (110.1 <= value && value <= 120) {
                 return 12;
             }
         }
-        return 0;
+        return 12;
 
     }
 
@@ -213,8 +243,7 @@ public class MosquitoFragment extends Fragment{
         NetworkInfo activeNetwork = cm.getActiveNetworkInfo();
         if (activeNetwork != null) { // connected to the internet
             if (activeNetwork.getType() == ConnectivityManager.TYPE_WIFI) {
-            }
-            else if (activeNetwork.getType() == ConnectivityManager.TYPE_MOBILE) {
+            } else if (activeNetwork.getType() == ConnectivityManager.TYPE_MOBILE) {
             }
             return true;
         } else {
@@ -222,30 +251,29 @@ public class MosquitoFragment extends Fragment{
         }
     }
 
-
     //값 불러오기
-     private boolean getPreferences(){
-         SharedPreferences pref = getContext().getSharedPreferences("pref", Context.MODE_PRIVATE);
-         value=pref.getString("value", "");
-         if(value == "")
-             return false;
-         date=pref.getString("dateView", "");
-         textView2.setText(date);
-         textView3.setText(value);
-         textView4.setText(getGrade(value));
-         return true;
-     }
+    private boolean getPreferences() {
+        SharedPreferences pref = getContext().getSharedPreferences("pref", Context.MODE_PRIVATE);
+        value = pref.getString("valueTextView", "");
+        if (value == "")
+            return false;
+        date = pref.getString("dateView", "");
+        textView2.setText(date);
+        textView3.setText(value);
+        textView4.setText(getGrade(value));
+        return true;
+    }
 
-     // 값 저장하기
-     private void savePreferences(){
-         SharedPreferences pref = getContext().getSharedPreferences("pref", Context.MODE_PRIVATE);
-         SharedPreferences.Editor editor = pref.edit();
-         editor.putString("value", value);
-         editor.putString("dateView", date);
-         editor.putInt("color", textColor);
-         editor.putString("grade", grade);
-         editor.commit();
-     }
+    // 값 저장하기
+    private void savePreferences() {
+        SharedPreferences pref = getContext().getSharedPreferences("pref", Context.MODE_PRIVATE);
+        SharedPreferences.Editor editor = pref.edit();
+        editor.putString("valueTextView", value);
+        editor.putString("dateView", date);
+        editor.putInt("color", textColor);
+        editor.putString("grade", grade);
+        editor.commit();
+    }
 
 
     private void setColor() {
@@ -259,27 +287,23 @@ public class MosquitoFragment extends Fragment{
 
     public String getGrade(String data) {
         double value = Double.valueOf(data);
-        if(value<=250){
-            textColor = Color.rgb(255,205,189);
+        if (value <= 30) {
+            textColor = Color.rgb(255, 205, 189);
             return "1단계(쾌적)";
-        }
-        else if(250.1<=value&&value<=500){
-            textColor = Color.rgb(255,153,51);
+        } else if (30.1 <= value && value <= 60) {
+            textColor = Color.rgb(255, 153, 51);
             return "2단계(관심)";
-        }
-        else if(500.1<=value&&value<=750){
-            textColor = Color.rgb(255,178,102);
+        } else if (60.1 <= value && value <= 90) {
+            textColor = Color.rgb(255, 178, 102);
             return "3단계(주의)";
-        }
-        else if(750.1<=value&&value<=1000){
+        } else if (90.1 <= value && value <= 120) {
             textColor = Color.RED;
             return "4단계(불쾌)";
-        }
-        else
-            return data;
+        } else
+            return "4단계(불쾌)";
     }
 
-    public String getXmlData(String date) {
+    public String getXmlData(String date, String criteria) {
         StringBuffer buffer = new StringBuffer();
 
         String queryUrl = "http://openapi.seoul.go.kr:8088/"
@@ -287,7 +311,7 @@ public class MosquitoFragment extends Fragment{
                 + "/xml"
                 + "/MosquitoStatus"
                 + "/1/5"
-                + "/"+date;
+                + "/" + date;
 
         try {
             URL url = new URL(queryUrl); //문자열로 된 요청 url을 URL 객체로 생성.
@@ -306,7 +330,8 @@ public class MosquitoFragment extends Fragment{
 
                     case XmlPullParser.START_TAG:
                         tag = xpp.getName();    //태그 이름 얻어오기
-                        if (tag.equals("MOSQUITO_VALUE")) {
+
+                        if (tag.equals(criteria)) {
                             xpp.next();
                             buffer.append(xpp.getText()); //grade 요소의 TEXT 읽어와서 문자열버퍼에 추가
                             buffer.append("\n");          //줄바꿈 문자 추가
@@ -316,7 +341,8 @@ public class MosquitoFragment extends Fragment{
 
                     case XmlPullParser.END_TAG:
                         tag = xpp.getName();    //태그 이름 얻어오기
-                        if (tag.equals("MOSQUITO_VALUE")) {
+
+                        if (tag.equals(criteria)) {
                             buffer.append("\n"); // 첫번째 검색결과종료..줄바꿈
                             break label1;
                         }
@@ -325,28 +351,27 @@ public class MosquitoFragment extends Fragment{
                 eventType = xpp.next();
 
             }
-            if(buffer.toString().trim().equals("")){
+            if (buffer.toString().trim().equals("")) {
                 throw new Exception("");
             }
         } catch (Exception e) {
-            return getXmlData(getYesterday());
+            return getXmlData(getYesterday(), "MOSQUITO_VALUE_HOUSE");
         }
         return buffer.toString().trim();
     }
 
 
-    public String getDateString()
-    {
+    public String getDateString() {
         SimpleDateFormat df = new SimpleDateFormat("yyyy-MM-dd", Locale.KOREA); // HH:mm:ss
         return df.format(new Date());
     }
 
     public String getDateStringToShow() {
-        SimpleDateFormat df = new SimpleDateFormat("yyyy.MM.dd("+getDay()+")", Locale.KOREA);
+        SimpleDateFormat df = new SimpleDateFormat("yyyy.MM.dd(" + getDay() + ")", Locale.KOREA);
         return df.format(new Date());
     }
 
-    public String getDay(){
+    public String getDay() {
         String day = "";
         Calendar cal = Calendar.getInstance();
         cal.setTime(new Date());
@@ -376,6 +401,7 @@ public class MosquitoFragment extends Fragment{
         }
         return day;
     }
+
     public String getYesterday() {
         SimpleDateFormat df = new SimpleDateFormat("yyyy-MM-dd", Locale.KOREA);
         return df.format(yesterday());
@@ -391,7 +417,6 @@ public class MosquitoFragment extends Fragment{
         super.onDestroy();
         savePreferences();
     }
-
 
 
 }
