@@ -2,7 +2,7 @@ package kr.cds.jisulife;
 
 import android.app.Application;
 
-public class ApplicationBase extends Application {      //폰트설정
+public class ApplicationBase extends Application {
     public void onCreate(){
         super.onCreate();
 
