@@ -1,6 +1,6 @@
 package kr.cds.jisulife.lifeIndex;
 import android.graphics.Color;
-import kr.cds.jisulife.LifeIndex;
+
 public class WinterLifeIndex extends LifeIndex {
 
     private int value;
